requirements.txt
# Основные зависимости для YOLO проектов
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
PyQt5>=5.15.0
Pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.64.0
PyYAML>=6.0
requests>=2.28.0
pathlib2>=2.3.0

# Для аугментации данных (yolo_train_au.py)
albumentations>=1.3.0

# Для анализа изображений (video.py)
scikit-image>=0.19.0

# Дополнительные зависимости для работы с изображениями
matplotlib>=3.5.0
scipy>=1.9.0

# Для работы с JSON и API
urllib3>=1.26.0

# Опциональные зависимости для ускорения работы
# torch>=2.0.0  # Раскомментируйте если нужна GPU поддержка
# torchvision>=0.15.0  # Раскомментируйте если нужна GPU поддержка